fix(dashboard): make protection repair failures actionable
