+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/app-detail-workspace.js
@@ -1,4 +1,4 @@
-import { r as reactExports, j as jsxRuntimeExports, aR as fetchApprovalPage, aS as fetchPolicy, v as guardActionDisposition, p as protectionHealthFor, k as useProtectionPresentationState, aN as HiMiniArrowLeft, c as HiMiniChevronRight, i as harnessDisplayName, q as GuardHero, Y as ProofStrip, aT as HiMiniHome, N as HiMiniBolt, a8 as HiMiniAdjustmentsHorizontal, S as SectionLabel, A as ActionButton, t as HiMiniShieldCheck, w as formatRelativeTime, M as HiMiniExclamationTriangle, aU as guardActionPresentation, P as Badge, aV as DEFAULT_FILTER_STATE, aW as filterEvidence, aX as sortEvidence, aY as computeMetrics, aZ as CommandActivityWorkspace, m as EmptyState, a_ as EvidenceFilterBar, a$ as EvidenceInsightStrip, b0 as EvidenceActionList, b1 as EvidenceActionDetail, K as useFocusTrap, b2 as policyIdentityKey, I as HiMiniCloud, b3 as HiMiniChartBar, av as Tag, o as HiMiniCheckCircle, _ as HiMiniXCircle, b4 as runHarnessAction, b5 as GuardHarnessActionError, b6 as HiMiniRocketLaunch, aA as HiMiniArrowPath, b7 as HiMiniTrash, b8 as clearLabelForScope, b9 as formatHarnessCommand } from "../guard-dashboard.js";
+import { r as reactExports, j as jsxRuntimeExports, aP as fetchApprovalPage, aQ as fetchPolicy, v as guardActionDisposition, p as protectionHealthFor, k as useProtectionPresentationState, aL as HiMiniArrowLeft, c as HiMiniChevronRight, i as harnessDisplayName, q as GuardHero, R as ProofStrip, aR as HiMiniHome, N as HiMiniBolt, a6 as HiMiniAdjustmentsHorizontal, S as SectionLabel, A as ActionButton, t as HiMiniShieldCheck, w as formatRelativeTime, M as HiMiniExclamationTriangle, aS as guardActionPresentation, P as Badge, aT as DEFAULT_FILTER_STATE, aU as filterEvidence, aV as sortEvidence, aW as computeMetrics, aX as CommandActivityWorkspace, m as EmptyState, aY as EvidenceFilterBar, aZ as EvidenceInsightStrip, a_ as EvidenceActionList, a$ as EvidenceActionDetail, K as useFocusTrap, b0 as policyIdentityKey, I as HiMiniCloud, b1 as HiMiniChartBar, at as Tag, o as HiMiniCheckCircle, X as HiMiniXCircle, b2 as runHarnessAction, b3 as GuardHarnessActionError, b4 as HiMiniRocketLaunch, ay as HiMiniArrowPath, b5 as HiMiniTrash, b6 as clearLabelForScope, b7 as formatHarnessCommand } from "../guard-dashboard.js";
 import { a as appSetupTarget } from "./harness-setup-target.js";
 function ActivityModeButton(props) {
   const active = props.mode === props.value;
