+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/extensions-workspace.js
@@ -1,4 +1,4 @@
-import { ax as fetchLocalCliApi, r as reactExports, ay as fetchExtensionControlApi, j as jsxRuntimeExports, az as useResolvedApprovalGate, a6 as HiMiniLockClosed, M as HiMiniExclamationTriangle, aA as HiMiniArrowPath, t as HiMiniShieldCheck, aB as HiMiniInformationCircle, aC as isApprovalProofSubmitDisabled, z as HiMiniXMark, aD as ApprovalProofFieldInputs, aE as buildApprovalProofCredentials, aF as GenIcon, N as HiMiniBolt, aG as HiMiniGlobeAlt, aH as HiMiniCube, I as HiMiniCloud, aI as HiMiniServerStack, b as HiMiniCommandLine, aJ as HiMiniFolder, aK as FaWindows, aL as FaAws, o as HiMiniCheckCircle, c as HiMiniChevronRight, C as HiMiniChevronDown, aM as approvalProofRecentlySatisfied, aN as HiMiniArrowLeft, aO as HiMiniPlus, $ as HiMiniClipboardDocumentCheck, a0 as HiMiniClipboard, as as HiMiniMagnifyingGlass, y as HiMiniSparkles, aP as HiMiniNoSymbol, ar as WorkspacePageHeader, aQ as guardAwareHref } from "../guard-dashboard.js";
+import { av as fetchLocalCliApi, r as reactExports, aw as fetchExtensionControlApi, j as jsxRuntimeExports, ax as useResolvedApprovalGate, a4 as HiMiniLockClosed, M as HiMiniExclamationTriangle, ay as HiMiniArrowPath, t as HiMiniShieldCheck, az as HiMiniInformationCircle, aA as isApprovalProofSubmitDisabled, z as HiMiniXMark, aB as ApprovalProofFieldInputs, aC as buildApprovalProofCredentials, aD as GenIcon, N as HiMiniBolt, aE as HiMiniGlobeAlt, aF as HiMiniCube, I as HiMiniCloud, aG as HiMiniServerStack, b as HiMiniCommandLine, aH as HiMiniFolder, aI as FaWindows, aJ as FaAws, o as HiMiniCheckCircle, c as HiMiniChevronRight, C as HiMiniChevronDown, aK as approvalProofRecentlySatisfied, aL as HiMiniArrowLeft, aM as HiMiniPlus, Z as HiMiniClipboardDocumentCheck, _ as HiMiniClipboard, aq as HiMiniMagnifyingGlass, y as HiMiniSparkles, aN as HiMiniNoSymbol, ap as WorkspacePageHeader, aO as guardAwareHref } from "../guard-dashboard.js";
 import { A as ApprovalProofModal } from "./approval-proof-modal.js";
 const EXTENSION_ID_PATTERN = /^command\.[a-z0-9]+(?:[.-][a-z0-9]+)*$/;
 const RULE_ID_PATTERN = /^command\.[a-z0-9]+(?:[.-][a-z0-9]+)*$/;
