+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/home-protection-module.js
@@ -1,4 +1,4 @@
-import { r as reactExports, j as jsxRuntimeExports, t as HiMiniShieldCheck, aB as HiMiniInformationCircle, M as HiMiniExclamationTriangle, S as SectionLabel, A as ActionButton, bN as HiMiniArrowRight, av as Tag, w as formatRelativeTime, B as HiMiniChevronUp, C as HiMiniChevronDown, o as HiMiniCheckCircle, _ as HiMiniXCircle } from "../guard-dashboard.js";
+import { r as reactExports, j as jsxRuntimeExports, t as HiMiniShieldCheck, az as HiMiniInformationCircle, M as HiMiniExclamationTriangle, S as SectionLabel, A as ActionButton, bM as HiMiniArrowRight, at as Tag, w as formatRelativeTime, B as HiMiniChevronUp, C as HiMiniChevronDown, o as HiMiniCheckCircle, X as HiMiniXCircle } from "../guard-dashboard.js";
 function resolveHomeProtectionStatus(snapshot) {
   const protection = snapshot.supply_chain?.package_manager_protection;
   if (!protection) return "unknown";
