+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/policy-workspace-page.js
@@ -1,4 +1,4 @@
-import { j as jsxRuntimeExports, S as SectionLabel, z as HiMiniXMark, P as Badge, av as Tag, b as HiMiniCommandLine, M as HiMiniExclamationTriangle, bD as scopeLabel, i as harnessDisplayName, A as ActionButton, aQ as guardAwareHref, w as formatRelativeTime$1, bE as HiMiniDocumentText, o as HiMiniCheckCircle, bF as HiMiniCloudArrowUp, bG as HiMiniCheck, bH as HiMiniCodeBracket, bI as HiMiniClipboardDocument, bJ as HiMiniUsers, bg as HiMiniBeaker, aJ as HiMiniFolder, a6 as HiMiniLockClosed, t as HiMiniShieldCheck, aB as HiMiniInformationCircle, bo as HiMiniCloudArrowDown, bn as HiMiniArrowTopRightOnSquare, bK as HiMiniIdentification, bL as policyActionLabel, r as reactExports, bM as createCloudExceptionRequest, bN as HiMiniArrowRight, m as EmptyState, as as HiMiniMagnifyingGlass, B as HiMiniChevronUp, C as HiMiniChevronDown, c as HiMiniChevronRight, bO as HiMiniPuzzlePiece, aG as HiMiniGlobeAlt, be as HiMiniClock, bP as fetchCloudExceptions, bQ as fetchCloudExceptionRequests, bR as downloadBlob, bS as PolicyStatField, bT as PaginationControls, aP as HiMiniNoSymbol, aH as HiMiniCube, aA as HiMiniArrowPath, I as HiMiniCloud, a8 as HiMiniAdjustmentsHorizontal, bU as HiMiniArrowDownTray, bV as HiMiniQueueList, ar as WorkspacePageHeader, bB as lazyWorkspace, bC as __vitePreload } from "../guard-dashboard.js";
+import { j as jsxRuntimeExports, S as SectionLabel, z as HiMiniXMark, P as Badge, at as Tag, b as HiMiniCommandLine, M as HiMiniExclamationTriangle, bC as scopeLabel, i as harnessDisplayName, A as ActionButton, aO as guardAwareHref, w as formatRelativeTime$1, bD as HiMiniDocumentText, o as HiMiniCheckCircle, bE as HiMiniCloudArrowUp, bF as HiMiniCheck, bG as HiMiniCodeBracket, bH as HiMiniClipboardDocument, bI as HiMiniUsers, be as HiMiniBeaker, aH as HiMiniFolder, a4 as HiMiniLockClosed, t as HiMiniShieldCheck, az as HiMiniInformationCircle, bm as HiMiniCloudArrowDown, bl as HiMiniArrowTopRightOnSquare, bJ as HiMiniIdentification, bK as policyActionLabel, r as reactExports, bL as createCloudExceptionRequest, bM as HiMiniArrowRight, m as EmptyState, aq as HiMiniMagnifyingGlass, B as HiMiniChevronUp, C as HiMiniChevronDown, c as HiMiniChevronRight, bN as HiMiniPuzzlePiece, aE as HiMiniGlobeAlt, bc as HiMiniClock, bO as fetchCloudExceptions, bP as fetchCloudExceptionRequests, bQ as downloadBlob, bR as PolicyStatField, bS as PaginationControls, aN as HiMiniNoSymbol, aF as HiMiniCube, ay as HiMiniArrowPath, I as HiMiniCloud, a6 as HiMiniAdjustmentsHorizontal, bT as HiMiniArrowDownTray, bU as HiMiniQueueList, ap as WorkspacePageHeader, bA as lazyWorkspace, bB as __vitePreload } from "../guard-dashboard.js";
 const CLOUD_EXCEPTION_EXPIRING_SOON_DAYS = 7;
 function parseCloudExceptionTimestamp(value) {
   if (!value || !value.trim()) {
