+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/settings-workspace.js
@@ -1,4 +1,4 @@
-import { a1 as PROTECTION_POSTURE_COPY, a2 as POSTURE_OUTCOME_COLUMNS, j as jsxRuntimeExports, r as reactExports, a3 as getDefaultExportFromCjs, a4 as React, K as useFocusTrap, a5 as HiMiniKey, S as SectionLabel, A as ActionButton, t as HiMiniShieldCheck, a6 as HiMiniLockClosed, a7 as HiMiniBellAlert, a8 as HiMiniAdjustmentsHorizontal, a9 as HiMiniCircleStack, aa as TabBar, c as HiMiniChevronRight, ab as resolveProtectionLevelCopy, ac as fetchSettings, ad as fetchRuntimeSnapshot, e as updateSettings, ae as clearPolicy, af as clearReviewQueue, ag as revokeApprovalGateCooldown, ah as disableApprovalGateTotp, ai as importSettings, aj as resetSettings, ak as enrollApprovalGateTotp, al as verifyApprovalGateTotp, am as clearEvidence, an as exportDiagnostics, ao as repairApprovalCenter, ap as exportSettings, aq as setupDesktopNotifications, m as EmptyState, ar as WorkspacePageHeader, W as WatchProtectionBanner, as as HiMiniMagnifyingGlass, C as HiMiniChevronDown, o as HiMiniCheckCircle, M as HiMiniExclamationTriangle, at as isProtectionPosture, au as deriveProtectionPosture, av as Tag, aw as approvalGateCooldownLabel, z as HiMiniXMark } from "../guard-dashboard.js";
+import { $ as PROTECTION_POSTURE_COPY, a0 as POSTURE_OUTCOME_COLUMNS, j as jsxRuntimeExports, r as reactExports, a1 as getDefaultExportFromCjs, a2 as React, K as useFocusTrap, a3 as HiMiniKey, S as SectionLabel, A as ActionButton, t as HiMiniShieldCheck, a4 as HiMiniLockClosed, a5 as HiMiniBellAlert, a6 as HiMiniAdjustmentsHorizontal, a7 as HiMiniCircleStack, a8 as TabBar, c as HiMiniChevronRight, a9 as resolveProtectionLevelCopy, aa as fetchSettings, ab as fetchRuntimeSnapshot, e as updateSettings, ac as clearPolicy, ad as clearReviewQueue, ae as revokeApprovalGateCooldown, af as disableApprovalGateTotp, ag as importSettings, ah as resetSettings, ai as enrollApprovalGateTotp, aj as verifyApprovalGateTotp, ak as clearEvidence, al as exportDiagnostics, am as repairApprovalCenter, an as exportSettings, ao as setupDesktopNotifications, m as EmptyState, ap as WorkspacePageHeader, W as WatchProtectionBanner, aq as HiMiniMagnifyingGlass, C as HiMiniChevronDown, o as HiMiniCheckCircle, M as HiMiniExclamationTriangle, ar as isProtectionPosture, as as deriveProtectionPosture, at as Tag, au as approvalGateCooldownLabel, z as HiMiniXMark } from "../guard-dashboard.js";
 import { f as filterSettingsBySearch, R as RISK_CONTROL_CONSEQUENCES } from "./app-catalog.js";
 const POSTURE_ORDER = ["protected", "extra_careful", "watch"];
 function ProtectionPosturePanel(props) {
