+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/supply-chain-hub-workspace.js
@@ -1,5 +1,5 @@
 const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/chunks/supply-chain-workspace.js","assets/guard-dashboard.js","assets/index.css","assets/chunks/feed-health-workspace.js","assets/chunks/home-protection-module.js","assets/chunks/supply-chain-protection-stats.js","assets/chunks/approval-proof-modal.js","assets/chunks/audit-workspace.js"])))=>i.map(i=>d[i]);
-import { ba as isSupplyChainAuditIncomplete, bb as isSupplyChainAuditEvidence, b5 as GuardHarnessActionError, bc as readString$1, bd as isRecord$1, r as reactExports, j as jsxRuntimeExports, o as HiMiniCheckCircle, aA as HiMiniArrowPath, M as HiMiniExclamationTriangle, av as Tag, w as formatRelativeTime, be as HiMiniClock, bf as IconActionButton, _ as HiMiniXCircle, b7 as HiMiniTrash, t as HiMiniShieldCheck, V as HiMiniWrenchScrewdriver, bg as HiMiniBeaker, bh as ActivationSummary, bi as ActionResultPanel, as as HiMiniMagnifyingGlass, m as EmptyState, A as ActionButton, bj as HiMiniBugAnt, z as HiMiniXMark, aE as buildApprovalProofCredentials, bk as GuardModalLayer, bl as ConnectFlowCard, bm as ApprovalProofInline, bn as HiMiniArrowTopRightOnSquare, bo as HiMiniCloudArrowDown, az as useResolvedApprovalGate, bp as fetchPackageFirewallStatus, bq as runPackageAudit, br as resolveSupplyChainAuditFailure, bs as runPackageSync, bt as startPackageFirewallConnect, U as openPackageFirewallAuthorizeFallback, bu as PACKAGE_FIREWALL_CONNECT_POPUP_BLOCKED_MESSAGE, bv as repairSupplyChainProtection, bw as runPackageFirewallAction, bx as parseInterceptProofSnapshot, by as activatePackageFirewallRuntime, S as SectionLabel, bz as EntitlementNotice, bA as fetchReceipts, ar as WorkspacePageHeader, bB as lazyWorkspace, bC as __vitePreload } from "../guard-dashboard.js";
+import { b8 as isSupplyChainAuditIncomplete, b9 as isSupplyChainAuditEvidence, b3 as GuardHarnessActionError, ba as readString$1, bb as isRecord$1, r as reactExports, j as jsxRuntimeExports, o as HiMiniCheckCircle, ay as HiMiniArrowPath, M as HiMiniExclamationTriangle, at as Tag, w as formatRelativeTime, bc as HiMiniClock, bd as IconActionButton, X as HiMiniXCircle, b5 as HiMiniTrash, t as HiMiniShieldCheck, V as HiMiniWrenchScrewdriver, be as HiMiniBeaker, bf as ActivationSummary, bg as ActionResultPanel, aq as HiMiniMagnifyingGlass, m as EmptyState, A as ActionButton, bh as HiMiniBugAnt, z as HiMiniXMark, aC as buildApprovalProofCredentials, bi as GuardModalLayer, bj as ConnectFlowCard, bk as ApprovalProofInline, bl as HiMiniArrowTopRightOnSquare, bm as HiMiniCloudArrowDown, ax as useResolvedApprovalGate, bn as fetchPackageFirewallStatus, bo as runPackageAudit, bp as resolveSupplyChainAuditFailure, bq as runPackageSync, br as startPackageFirewallConnect, bs as openPackageFirewallAuthorizeFallback, bt as PACKAGE_FIREWALL_CONNECT_POPUP_BLOCKED_MESSAGE, bu as repairSupplyChainProtection, bv as runPackageFirewallAction, bw as parseInterceptProofSnapshot, bx as activatePackageFirewallRuntime, S as SectionLabel, by as EntitlementNotice, bz as fetchReceipts, ap as WorkspacePageHeader, bA as lazyWorkspace, bB as __vitePreload } from "../guard-dashboard.js";
 import { A as ApprovalProofModal } from "./approval-proof-modal.js";
 const SEVERITY_RANK = {
   critical: 4,
